getXYZfromMDA look like it works
