+++ pyproject.toml
@@ -51,7 +51,7 @@ extra-dependencies = [
 
 [tool.hatch.envs.test.scripts]
 cov = 'pytest --cov-report=term-missing --cov-config=pyproject.toml --cov=pkg --cov=tests'
-
+tools = 'pytest ./tests/test_tools.py'
 
 [[tool.hatch.envs.test.matrix]]
 python = ["3.8", "3.9"]
